Consolidate common rpm-ostree and bootc functions into image_manager_helper.go
