+++ pkg/daemon/rpm-ostree.go
@@ -5,43 +5,13 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
-	"time"
 
 	rpmostreeclient "github.com/coreos/rpmostree-client-go/pkg/client"
-	"github.com/opencontainers/go-digest"
-	pivotutils "github.com/openshift/machine-config-operator/pkg/daemon/pivot/utils"
 	"gopkg.in/yaml.v2"
 	"k8s.io/klog/v2"
 )
-
-const (
-	// the number of times to retry commands that pull data from the network
-	numRetriesNetCommands = 5
-	// Default ostreeAuthFile location
-	ostreeAuthFile = "/run/ostree/auth.json"
-	// Pull secret.  Written by the machine-config-operator
-	kubeletAuthFile = "/var/lib/kubelet/config.json"
-	// Internal Registry Pull secret + Global Pull secret.  Written by the machine-config-operator.
-	imageRegistryAuthFile = "/etc/mco/internal-registry-pull-secret.json"
-)
-
-// imageInspection is a public implementation of
-// https://github.com/containers/skopeo/blob/82186b916faa9c8c70cfa922229bafe5ae024dec/cmd/skopeo/inspect.go#L20-L31
-type imageInspection struct {
-	Name          string `json:",omitempty"`
-	Tag           string `json:",omitempty"`
-	Digest        digest.Digest
-	RepoDigests   []string
-	Created       *time.Time
-	DockerVersion string
-	Labels        map[string]string
-	Architecture  string
-	Os            string
-	Layers        []string
-}
 
 // RpmOstreeClient provides all RpmOstree related methods in one structure.
 // This structure implements DeploymentClient
@@ -163,38 +133,6 @@ func (r *RpmOstreeClient) GetBootedOSImageURL() (string, string, string, error) 
 	return osImageURL, bootedDeployment.Version, baseChecksum, nil
 }
 
-func podmanInspect(imgURL string) (imgdata *imageInspection, err error) {
-	// Pull the container image if not already available
-	var authArgs []string
-	if _, err := os.Stat(ostreeAuthFile); err == nil {
-		authArgs = append(authArgs, "--authfile", ostreeAuthFile)
-	}
-	args := []string{"pull", "-q"}
-	args = append(args, authArgs...)
-	args = append(args, imgURL)
-	_, err = pivotutils.RunExt(numRetriesNetCommands, "podman", args...)
-	if err != nil {
-		return
-	}
-
-	inspectArgs := []string{"inspect", "--type=image"}
-	inspectArgs = append(inspectArgs, fmt.Sprintf("%s", imgURL))
-	var output []byte
-	output, err = runGetOut("podman", inspectArgs...)
-	if err != nil {
-		return
-	}
-	var imagedataArray []imageInspection
-	err = json.Unmarshal(output, &imagedataArray)
-	if err != nil {
-		err = fmt.Errorf("unmarshaling podman inspect: %w", err)
-		return
-	}
-	imgdata = &imagedataArray[0]
-	return
-
-}
-
 // RpmOstreeIsNewEnoughForLayering returns true if the version of rpm-ostree on the
 // host system is new enough for layering.
 // VersionData represents the static information about rpm-ostree.
@@ -296,33 +234,4 @@ func useMergedPullSecrets() error {
 
 	// Attempt to link the merged secrets file
 	return linkOstreeAuthFile(imageRegistryAuthFile)
-}
-
-// truncate a string using runes/codepoints as limits.
-// This specifically will avoid breaking a UTF-8 value.
-func truncate(input string, limit int) string {
-	asRunes := []rune(input)
-	l := len(asRunes)
-
-	if limit >= l {
-		return input
-	}
-
-	return fmt.Sprintf("%s [%d more chars]", string(asRunes[:limit]), l-limit)
-}
-
-// runGetOut executes a command, logging it, and return the stdout output.
-func runGetOut(command string, args ...string) ([]byte, error) {
-	klog.Infof("Running captured: %s %s", command, strings.Join(args, " "))
-	cmd := exec.Command(command, args...)
-	rawOut, err := cmd.Output()
-	if err != nil {
-		errtext := ""
-		if e, ok := err.(*exec.ExitError); ok {
-			// Trim to max of 256 characters
-			errtext = fmt.Sprintf("\n%s", truncate(string(e.Stderr), 256))
-		}
-		return nil, fmt.Errorf("error running %s %s: %s%s", command, strings.Join(args, " "), err, errtext)
-	}
-	return rawOut, nil
 }
